// src/v1/modules/campusManagement/dtos/add-campus.dto.ts
export type AddCampus = {
  campusName: string;
  campusCode: string;
  campusAddress: string;
  campusEmail: string;
  campusPhoneNumber: string;
  campusPastor: string;
  financeManager?: string;
  legalName?: string;
  taxId?: string;
  registrationNumber?: string;
  localCurrency: string;
  timezone: string;
  established: string;
  status: string;
  churchId: string;
};